pyproject.toml
[project]
name = "loud_bot"
version = "0.1.0"
description = "LOUD stands for Local Output Uploaded to Discord"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "discord-py[voice]>=2.4.0",
    "fastapi>=0.122.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.2.1",
    "types-yt-dlp>=2025.12.8.20251210",
    "uvicorn[standard]>=0.38.0",
    "yt-dlp>=2025.12.8",
]
licence = { text = "MIT" }

[tool.poe.tasks]
# Testing
test = "pytest -m 'not integration'"
test-integration = "pytest -m integration"
test-all = "pytest"
test-cov = "pytest --cov=src --cov-report=term-missing --cov-report=html"
test-watch = "pytest --watch"

# Linting and formatting
lint = "ruff check ."
lint-fix = "ruff check --fix ."
format = "ruff format ."
format-check = "ruff format --check ."

# Type checking
typecheck = "mypy src"

# Combined quality checks
check = ["format-check", "lint", "typecheck", "test"]

api = "python -m src.api"
bot = "python -m src.bot"

dev = ["api", "bot"]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "poethepoet>=0.38.0",
    "ruff>=0.14.7",
]
test = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
